--- src/components/Navigation/index.js
@@ -1,44 +1,0 @@
-import React from 'react';
-import { Link } from 'react-router-dom';
-
-import SignOutButton from '../SignOut';
-import * as ROUTES from '../../constants/routes';
-import { AuthUserContext } from '../Session';
-
-const Navigation = () => (
-  <AuthUserContext.Consumer>
-    {authUser =>
-      authUser ? <NavigationAuth /> : <NavigationNonAuth />
-    }
-  </AuthUserContext.Consumer>
-)
-
-const NavigationAuth = () => (
-  <ul>
-    <li>
-      <Link to={ROUTES.LANDING}>Landing</Link>
-    </li>
-    <li>
-      <Link to={ROUTES.MAIN}>Main</Link>
-    </li>
-    <li>
-      <Link to={ROUTES.ACCOUNT}>Account</Link>
-    </li>
-    <li>
-      <SignOutButton />
-    </li>
-  </ul>
-);
-
-const NavigationNonAuth = () => (
-  <ul>
-    <li>
-      <Link to={ROUTES.LANDING}>Landing</Link>
-    </li>
-    <li>
-      <Link to={ROUTES.SIGN_IN}>Sign In</Link>
-    </li>
-  </ul>
-);
-
-export default Navigation;
